Refactor Cypress workflows for GitHub Actions
The Cypress workflow has been split from `push_request.yml`, which has been deleted, to a new `cypress_push.yml` workflow. The new workflow now allows for the running of tests in parallel for better efficiency. Minor updates are also reflected in the ".idea/sonarlint" files and ".gitignore", and a new Project ID is added in `cypress.config.js`.